# spec/protocol.openapi.yaml
openapi: 3.1.0
info:
  title: Local Messaging Protocol
  version: "1"
  description: |
    Wire contract for the messaging server and both clients. Behavior rules are in
    spec/behavior.md; where this file cannot express a rule (UTF-8 byte limits,
    atomicity, cursor semantics), behavior.md is authoritative.

    Usernames on the wire MUST already be canonical (behavior.md section 2).
    Clients normalize before sending; the server rejects non-canonical names
    with INVALID_REQUEST rather than normalizing them.
servers:
  - url: http://localhost:8000
    description: Local development server (host and port are configurable)

paths:
  /v1/meta:
    get:
      operationId: getMeta
      summary: Protocol version and current server epoch
      responses:
        "200":
          description: Server metadata
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Meta" }

  /v1/messages:
    post:
      operationId: submitMessage
      summary: Accept a message, or return the original acceptance for an identical retry
      parameters:
        - $ref: "#/components/parameters/ServerEpoch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SubmitRequest" }
      responses:
        "200":
          description: Accepted (first time or identical retry; bodies are identical)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AcceptedMessage" }
        "409":
          description: MESSAGE_ID_CONFLICT or SERVER_EPOCH_CHANGED
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: INVALID_REQUEST
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  /v1/inbox/{username}:
    get:
      operationId: getInbox
      summary: Messages addressed to username with sequence greater than after
      parameters:
        - $ref: "#/components/parameters/ServerEpoch"
        - name: username
          in: path
          required: true
          schema: { $ref: "#/components/schemas/Username" }
        - name: after
          in: query
          required: false
          description: Exclusive cursor. Must not exceed the session's highest sequence.
          schema: { type: integer, format: int64, minimum: 0, default: 0 }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        "200":
          description: One page in ascending sequence order
          headers:
            Cache-Control:
              description: Always no-store
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InboxPage" }
        "409":
          description: SERVER_EPOCH_CHANGED
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: INVALID_REQUEST or INVALID_CURSOR
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

components:
  parameters:
    ServerEpoch:
      name: X-Server-Epoch
      in: header
      required: true
      description: Epoch saved from GET /v1/meta. Missing or empty is INVALID_REQUEST.
      schema: { type: string, minLength: 1, maxLength: 64 }

  schemas:
    Username:
      type: string
      pattern: "^[a-z0-9_]{1,32}$"

    MessageId:
      type: string
      description: Lowercase hyphenated UUID
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$"

    Text:
      type: string
      minLength: 1
      maxLength: 4096
      description: |
        Exact text. The real limit is 1-4096 UTF-8 bytes (behavior.md 3.2);
        maxLength here counts code points and is only a necessary condition.
        Never trimmed or normalized.

    Meta:
      type: object
      additionalProperties: true
      required: [protocol_version, server_epoch]
      properties:
        protocol_version: { type: string, const: "1" }
        server_epoch: { type: string, minLength: 1, maxLength: 64 }

    SubmitRequest:
      type: object
      additionalProperties: false
      required: [client_message_id, sender, recipient, text]
      properties:
        client_message_id: { $ref: "#/components/schemas/MessageId" }
        sender: { $ref: "#/components/schemas/Username" }
        recipient: { $ref: "#/components/schemas/Username" }
        text: { $ref: "#/components/schemas/Text" }

    AcceptedMessage:
      type: object
      additionalProperties: true
      required: [client_message_id, sender, recipient, text, sequence, accepted_at, server_epoch]
      properties:
        client_message_id: { $ref: "#/components/schemas/MessageId" }
        sender: { $ref: "#/components/schemas/Username" }
        recipient: { $ref: "#/components/schemas/Username" }
        text: { $ref: "#/components/schemas/Text" }
        sequence: { type: integer, format: int64, minimum: 1 }
        accepted_at:
          type: string
          description: UTC, millisecond precision, e.g. 2026-10-03T14:05:09.123Z
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$"
        server_epoch: { type: string, minLength: 1, maxLength: 64 }

    InboxPage:
      type: object
      additionalProperties: true
      required: [server_epoch, messages, next_cursor, has_more]
      properties:
        server_epoch: { type: string, minLength: 1, maxLength: 64 }
        messages:
          type: array
          maxItems: 100
          items: { $ref: "#/components/schemas/AcceptedMessage" }
        next_cursor: { type: integer, format: int64, minimum: 0 }
        has_more: { type: boolean }

    Error:
      type: object
      additionalProperties: true
      required: [code, message]
      properties:
        code:
          type: string
          enum: [INVALID_REQUEST, MESSAGE_ID_CONFLICT, SERVER_EPOCH_CHANGED, INVALID_CURSOR]
        message: { type: string }
        server_epoch: { type: [string, "null"] }